+++ packages/cli/tests/commands/collect/collect.test.ts
@@ -7,22 +7,42 @@ import {
   resetSetupSandbox,
   SETUP_SANDBOX_CLI_TEST_CFG,
   SETUP_SANDBOX_DEFAULT_PERSIST_OUT_PATH,
-  SETUP_SANDBOX_REMOTE_RC_NAME
+  SETUP_SANDBOX_REMOTE_RC_NAME, SETUP_SANDBOX_STATIC_RC_JSON, SETUP_SANDBOX_STATIC_RC_NAME
 } from '../../fixtures/setup-sandbox';
 import {
-  expectCollectCreatesJsonReport
+  expectCollectCreatesJsonReport, expectCollectNoLogsFromMockInStdout, expectCollectNotToCreateAReport
 } from '../../utils/cli-expectations';
 
 const defaultCommand = [CLI_PATH];
 const collectCommand = [...defaultCommand, 'collect', '-v'];
 const collectCommandRemoteRc = [...collectCommand, `-p=./${SETUP_SANDBOX_REMOTE_RC_NAME}`];
+const collectCommandStaticRc = [...collectCommand, `-p=./${SETUP_SANDBOX_STATIC_RC_NAME}`];
 
 const uf1Name = 'Sandbox Setup UF1';
 const uf1OutPathJons = path.join(SETUP_SANDBOX_DEFAULT_PERSIST_OUT_PATH, 'sandbox-setup-uf1.uf.json');
 
+const ufStaticName = 'Sandbox Setup StaticDist';
+const uf1OutPath = path.join(SETUP_SANDBOX_DEFAULT_PERSIST_OUT_PATH, 'sandbox-setup-uf1.uf.html');
+
 describe('collect command in setup sandbox', () => {
   beforeEach(async () => resetSetupSandbox());
-  afterEach(async () => resetSetupSandbox());
+  // afterEach(async () => resetSetupSandbox());
+
+
+  it('should load ufPath and execute the user-flow with verbose=false and save the report', async () => {
+    const { exitCode, stdout, stderr } = await cliPromptTest(
+      [...collectCommandStaticRc, `--ufPath=${SETUP_SANDBOX_STATIC_RC_JSON.collect.ufPath}`, '-v=false'],
+      [],
+      SETUP_SANDBOX_CLI_TEST_CFG
+    );
+
+    expect(stderr).toBe('');
+    expectCollectNoLogsFromMockInStdout(stdout, ufStaticName, SETUP_SANDBOX_STATIC_RC_JSON);
+    expect(exitCode).toBe(0);
+
+    expectCollectNotToCreateAReport(uf1OutPath);
+
+  }, 40_000);
 
 
   it('should load ufPath, execute the user-flow on a remote URL and save the file', async () => {
